feat: Move ProductShelf to app folder
