+++ packages/core/app/components/sections/ProductShelf/DefaultComponents.ts
