+++ packages/core/app/components/sections/ProductShelf/OverriddenDefaultProductShelf.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import { getOverriddenSection } from 'app/sdk/overrides/getOverriddenSection'
 import { override } from 'src/customizations/src/components/overrides/ProductShelf'
 import ProductShelf from '.'
