+++ packages/core/app/components/sections/ProductShelf/index.ts
