+++ packages/core/app/components/sections/ProductShelf/section.module.scss
