+++ packages/core/app/components/ui/ProductShelf/index.ts
