+++ packages/core/app/fs-next-update/page.tsx
@@ -5,6 +5,8 @@ import CUSTOM_COMPONENTS from 'src/customizations/src/components'
 import RenderSections from 'app/components/cms/RenderSections'
 import { OverriddenDefaultBannerText as BannerText } from 'app/components/sections/BannerText/OverriddenDefaultBannerText'
 import { OverriddenDefaultHero as Hero } from 'app/components/sections/Hero/OverriddenDefaultHero'
+import { OverriddenDefaultProductShelf as ProductShelf } from 'app/components/sections/ProductShelf/OverriddenDefaultProductShelf'
+
 import PageProvider from 'app/sdk/overrides/PageProvider'
 import { getDynamicContent } from 'app/utils/dynamicContent'
 import { ComponentType } from 'react'
@@ -13,6 +15,7 @@ import storeConfig from '../../faststore.config'
 /* A list of components that can be used in the CMS. */
 const COMPONENTS: Record<string, ComponentType<any>> = {
   Hero,
+  ProductShelf,
   BannerText,
   ...CUSTOM_COMPONENTS,
 }
