+++ packages/core/app/typings/overrides.ts
@@ -47,7 +47,7 @@ import EmptyState from '../../src/components/sections/EmptyState'
 import Newsletter from '../../src/components/sections/Newsletter'
 import ProductDetails from '../../src/components/sections/ProductDetails'
 import ProductGallery from '../../src/components/sections/ProductGallery'
-import ProductShelf from '../../src/components/sections/ProductShelf'
+import ProductShelf from '../components/sections/ProductShelf'
 import Alert from '../components/sections/Alert'
 import BannerText from '../components/sections/BannerText'
 import Hero from '../components/sections/Hero'
