+++ packages/core/src/components/sections/CrossSellingShelf/CrossSellingShelf.tsx
@@ -2,13 +2,13 @@
 
 import { useMemo } from 'react'
 
-import UIProductShelf from '../../../components/ui/ProductShelf'
+import UIProductShelf from 'app/components/ui/ProductShelf'
 import { useInView } from 'react-intersection-observer'
 import { usePDP } from 'app/sdk/overrides/PageProvider'
-import styles from '../ProductShelf/section.module.scss'
+import styles from 'app/components/sections/ProductShelf/section.module.scss'
 import Section from '../Section'
 import { CrossSellingShelfDefaultComponents } from './DefaultComponents'
-import { getOverridableSection } from '../../../../app/sdk/overrides/getOverriddenSection'
+import { getOverridableSection } from 'app/sdk/overrides/getOverriddenSection'
 
 interface Props {
   numberOfItems: number
